LicenseFilenamePatterns: Factor out getAllAncestorDirectories() to Utils
For later re-use.
